Move getRestrictedInt helpers into base/util (#4593)
