+++ base/util_test.go
@@ -1220,3 +1220,94 @@ func BenchmarkPanicRecover(b *testing.B) {
 		}
 	})
 }
+
+func TestGetRestrictedIntQuery(t *testing.T) {
+
+	defaultValue := uint64(42)
+	minValue := uint64(20)
+	maxValue := uint64(100)
+
+	// make sure it returns default value when passed empty Values
+	values := make(url.Values)
+	restricted := GetRestrictedIntQuery(
+		values,
+		"foo",
+		defaultValue,
+		minValue,
+		maxValue,
+		false,
+	)
+	assert.Equal(t, defaultValue, restricted)
+
+	// make sure it returns default value when passed Values that doesn't contain key
+	values.Set("bar", "99")
+	restricted = GetRestrictedIntQuery(
+		values,
+		"foo",
+		defaultValue,
+		minValue,
+		maxValue,
+		false,
+	)
+	assert.Equal(t, defaultValue, restricted)
+
+	// make sure it returns appropriate value from Values
+	values.Set("foo", "99")
+	restricted = GetRestrictedIntQuery(
+		values,
+		"foo",
+		defaultValue,
+		minValue,
+		maxValue,
+		false,
+	)
+	assert.Equal(t, uint64(99), restricted)
+
+	// make sure it is limited to max when value value is over max
+	values.Set("foo", "200")
+	restricted = GetRestrictedIntQuery(
+		values,
+		"foo",
+		defaultValue,
+		minValue,
+		maxValue,
+		false,
+	)
+	assert.Equal(t, maxValue, restricted)
+
+	// make sure it is limited to min when value value is under min
+	values.Set("foo", "1")
+	restricted = GetRestrictedIntQuery(
+		values,
+		"foo",
+		defaultValue,
+		minValue,
+		maxValue,
+		false,
+	)
+	assert.Equal(t, minValue, restricted)
+
+	// Return zero when allowZero=true
+	values.Set("foo", "0")
+	restricted = GetRestrictedIntQuery(
+		values,
+		"foo",
+		defaultValue,
+		minValue,
+		maxValue,
+		true,
+	)
+	assert.Equal(t, uint64(0), restricted)
+
+	// Return minValue when allowZero=false
+	values.Set("foo", "0")
+	restricted = GetRestrictedIntQuery(
+		values,
+		"foo",
+		defaultValue,
+		minValue,
+		maxValue,
+		false,
+	)
+	assert.Equal(t, minValue, restricted)
+}
