+++ rest/handler_test.go
@@ -2,102 +2,10 @@ package rest
 
 import (
 	"net/http"
-	"net/url"
 	"testing"
 
 	goassert "github.com/couchbaselabs/go.assert"
 )
-
-func TestGetRestrictedIntQuery(t *testing.T) {
-
-	defaultValue := uint64(42)
-	minValue := uint64(20)
-	maxValue := uint64(100)
-
-	// make sure it returns default value when passed empty Values
-	values := make(url.Values)
-	restricted := getRestrictedIntQuery(
-		values,
-		"foo",
-		defaultValue,
-		minValue,
-		maxValue,
-		false,
-	)
-	goassert.Equals(t, restricted, defaultValue)
-
-	// make sure it returns default value when passed Values that doesn't contain key
-	values.Set("bar", "99")
-	restricted = getRestrictedIntQuery(
-		values,
-		"foo",
-		defaultValue,
-		minValue,
-		maxValue,
-		false,
-	)
-	goassert.Equals(t, restricted, defaultValue)
-
-	// make sure it returns appropriate value from Values
-	values.Set("foo", "99")
-	restricted = getRestrictedIntQuery(
-		values,
-		"foo",
-		defaultValue,
-		minValue,
-		maxValue,
-		false,
-	)
-	goassert.Equals(t, restricted, uint64(99))
-
-	// make sure it is limited to max when value value is over max
-	values.Set("foo", "200")
-	restricted = getRestrictedIntQuery(
-		values,
-		"foo",
-		defaultValue,
-		minValue,
-		maxValue,
-		false,
-	)
-	goassert.Equals(t, restricted, maxValue)
-
-	// make sure it is limited to min when value value is under min
-	values.Set("foo", "1")
-	restricted = getRestrictedIntQuery(
-		values,
-		"foo",
-		defaultValue,
-		minValue,
-		maxValue,
-		false,
-	)
-	goassert.Equals(t, restricted, minValue)
-
-	// Return zero when allowZero=true
-	values.Set("foo", "0")
-	restricted = getRestrictedIntQuery(
-		values,
-		"foo",
-		defaultValue,
-		minValue,
-		maxValue,
-		true,
-	)
-	goassert.Equals(t, restricted, uint64(0))
-
-	// Return minValue when allowZero=false
-	values.Set("foo", "0")
-	restricted = getRestrictedIntQuery(
-		values,
-		"foo",
-		defaultValue,
-		minValue,
-		maxValue,
-		false,
-	)
-	goassert.Equals(t, restricted, minValue)
-}
 
 func TestParseHTTPRangeHeader(t *testing.T) {
 	// http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.35
